feat: add update source selection to network settings and update related descriptions
